+++ projects/rocfft/library/src/plan.cpp
@@ -542,6 +542,111 @@ catch(...)
     return rocfft_handle_exception();
 }
 
+static bool valid_symbol_name(const char* symbol_name)
+{
+    if(std::isdigit(symbol_name[0]))
+        return false;
+
+    static constexpr auto legal_chars = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ_";
+    constexpr auto legal_chars_end    = legal_chars + std::char_traits<char>::length(legal_chars);
+
+    const char* end = symbol_name + strlen(symbol_name);
+    return std::all_of(symbol_name, end, [=](char c) {
+        return c == '_' || std::isdigit(c)
+               || std::find(legal_chars, legal_chars_end, c) != legal_chars_end;
+    });
+}
+
+rocfft_status rocfft_plan_description_set_load_callback(rocfft_plan_description description,
+                                                        const char*             symbol_name,
+                                                        const void*             bitcode_data,
+                                                        size_t                  bitcode_len_bytes,
+                                                        void**                  cb_data,
+                                                        size_t                  shared_mem_bytes)
+try
+{
+    log_trace(__func__,
+              "description",
+              description,
+              "symbol_name",
+              symbol_name,
+              "bitcode_data",
+              bitcode_data,
+              "bitcode_len_bytes",
+              bitcode_len_bytes,
+              "cb_data",
+              cb_data,
+              "shared_mem_bytes",
+              shared_mem_bytes);
+    if(!description)
+        return rocfft_status_invalid_arg_value;
+
+    // shared memory for callbacks is not currently supported
+    if(shared_mem_bytes)
+        return rocfft_status_invalid_arg_value;
+
+    // clear the callback
+    if(!symbol_name || !bitcode_data || !bitcode_len_bytes)
+        description->loadOps.spirv_cb = {};
+    else
+    {
+        // validate that the symbol name is a legal C identifier
+        if(!valid_symbol_name(symbol_name))
+            return rocfft_status_invalid_arg_value;
+        description->loadOps.spirv_cb.set(symbol_name, bitcode_data, bitcode_len_bytes, cb_data);
+    }
+    return rocfft_status_success;
+}
+catch(...)
+{
+    return rocfft_handle_exception();
+}
+
+rocfft_status rocfft_plan_description_set_store_callback(rocfft_plan_description description,
+                                                         const char*             symbol_name,
+                                                         const void*             bitcode_data,
+                                                         size_t                  bitcode_len_bytes,
+                                                         void**                  cb_data,
+                                                         size_t                  shared_mem_bytes)
+try
+{
+    log_trace(__func__,
+              "description",
+              description,
+              "symbol_name",
+              symbol_name,
+              "bitcode_data",
+              bitcode_data,
+              "bitcode_len_bytes",
+              bitcode_len_bytes,
+              "cb_data",
+              cb_data,
+              "shared_mem_bytes",
+              shared_mem_bytes);
+    if(!description)
+        return rocfft_status_invalid_arg_value;
+
+    // shared memory for callbacks is not currently supported
+    if(shared_mem_bytes)
+        return rocfft_status_invalid_arg_value;
+
+    // clear the callback
+    if(!symbol_name || !bitcode_data || !bitcode_len_bytes)
+        description->storeOps.spirv_cb = {};
+    else
+    {
+        // validate that the symbol name is a legal C identifier
+        if(!valid_symbol_name(symbol_name))
+            return rocfft_status_invalid_arg_value;
+        description->storeOps.spirv_cb.set(symbol_name, bitcode_data, bitcode_len_bytes, cb_data);
+    }
+    return rocfft_status_success;
+}
+catch(...)
+{
+    return rocfft_handle_exception();
+}
+
 std::vector<size_t> rocfft_plan_t::get_user_facing_lengths() const
 {
     if(transformType == rocfft_transform_type_complex_forward